Prompt user to enter footprint name if unknown
Don't error if footprint is unknown (i.e. not SMT), just ask the user to
input footprint lib/name.

Don't perform heroics to figure out capacitor footprint name - if it's
not SMT, just ask the user. This deletes some mostly-working THT
capacitor footprint identification code, which maybe can be replaced
later, but it's not great to guess a footprint name if we aren't
checking that it exists in KiCad, and there's a lot of bad data out
there.